experiments/lr_research_demo/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
pyyaml==6.0
